chore: 패키지 구조 분리, api dto 네이밍 통일
