+++ backend/src/main/java/com/stampcrush/backend/api/customer/request/TemporaryCustomerCreateRequest.java
@@ -1,0 +1,15 @@
+package com.stampcrush.backend.api.customer.request;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+public class TemporaryCustomerCreateRequest {
+
+    @NotNull
+    private String phoneNumber;
+}
