+++ backend/src/main/java/com/stampcrush/backend/api/customer/response/CustomersFindResponse.java
@@ -1,0 +1,16 @@
+package com.stampcrush.backend.api.customer.response;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PUBLIC)
+public class CustomersFindResponse {
+
+    private List<CustomerFindResponse> customer;
+}
