--- backend/src/main/java/com/stampcrush/backend/api/dto/response/CafesFindResponse.java
@@ -1,13 +1,0 @@
-package com.stampcrush.backend.api.dto.response;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-import java.util.List;
-
-@Getter
-@RequiredArgsConstructor
-public class CafesFindResponse {
-
-    private final List<CafeFindResponse> cafes;
-}
