+++ backend/src/main/java/com/stampcrush/backend/api/reward/request/RewardUsedUpdateRequest.java
@@ -1,0 +1,15 @@
+package com.stampcrush.backend.api.reward.request;
+
+import jakarta.validation.constraints.AssertTrue;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class RewardUsedUpdateRequest {
+
+    private final Long cafeId;
+
+    @AssertTrue
+    private final boolean used;
+}
