+++ backend/src/main/java/com/stampcrush/backend/api/reward/response/RewardFindResponse.java
@@ -1,0 +1,18 @@
+package com.stampcrush.backend.api.reward.response;
+
+import com.stampcrush.backend.application.reward.dto.RewardFindResult;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class RewardFindResponse {
+
+    private final Long id;
+    private final String name;
+
+    public RewardFindResponse(RewardFindResult rewardFindResult) {
+        this.id = rewardFindResult.getId();
+        this.name = rewardFindResult.getName();
+    }
+}
