+++ backend/src/main/java/com/stampcrush/backend/api/reward/response/RewardsFindResponse.java
@@ -1,0 +1,13 @@
+package com.stampcrush.backend.api.reward.response;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+import java.util.List;
+
+@Getter
+@RequiredArgsConstructor
+public class RewardsFindResponse {
+
+    private final List<RewardFindResponse> rewards;
+}
