+++ backend/src/main/java/com/stampcrush/backend/api/sample/request/SampleCouponFindRequest.java
@@ -1,0 +1,11 @@
+package com.stampcrush.backend.api.sample.request;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class SampleCouponFindRequest {
+
+    private final Integer maxStampCount;
+}
