--- backend/src/main/java/com/stampcrush/backend/controller/request/SampleCouponQueryRequest.java
@@ -1,11 +1,0 @@
-package com.stampcrush.backend.controller.request;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class SampleCouponQueryRequest {
-
-    private final Integer maxStampCount;
-}
