--- backend/src/main/java/com/stampcrush/backend/service/CustomersResponse.java
@@ -1,16 +1,0 @@
-package com.stampcrush.backend.service;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PUBLIC)
-public class CustomersResponse {
-
-    private List<CustomerResponse> customer;
-}
