--- backend/src/main/java/com/stampcrush/backend/service/dto/CouponSettingDto.java
@@ -1,47 +1,0 @@
-package com.stampcrush.backend.service.dto;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.ToString;
-
-import java.util.List;
-
-@Getter
-@ToString
-@RequiredArgsConstructor
-public class CouponSettingDto {
-
-    private final CafeCouponDesignDto cafeCouponDesignDto;
-    private final CafePolicyDto cafePolicyDto;
-
-    @Getter
-    @ToString
-    @RequiredArgsConstructor
-    public static class CafeCouponDesignDto {
-
-        private final String frontImageUrl;
-        private final String backImageUrl;
-        private final String stampImageUrl;
-        private final List<CafeStampCoordinateDto> coordinates;
-
-        @Getter
-        @ToString
-        @RequiredArgsConstructor
-        public static class CafeStampCoordinateDto {
-
-            private final Integer order;
-            private final Integer xCoordinate;
-            private final Integer yCoordinate;
-        }
-    }
-
-    @Getter
-    @ToString
-    @RequiredArgsConstructor
-    public static class CafePolicyDto {
-
-        private final Integer maxStampCount;
-        private final String reward;
-        private final Integer expirePeriod;
-    }
-}
